Update code for games

+++ app/src/main/java/com/roxanasultan/memoraid/models/Level.kt
@@ -1,7 +1,7 @@
 package com.roxanasultan.memoraid.models
 
 data class Level(
-    val bestTime: Int = 0,
+    val bestTime: Long = 0,
     val lastPlayedGames: List<LastPlayedGame> = emptyList(),
     val leastMoves: Int = 0,
     val level: String = "",
